requirements.txt
streamlit==1.35.0
python-docx==1.1.0
PyMuPDF==1.24.3
